backend/requirements.txt
click==8.1.7
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3
protoletariat==3.3.2
